Move Path and PathMetrics from canvas.dart into their own files. No delta (#13772)
* Move Path and PathMetrics into their own files. No delta
* update file list in licenses_flutter
