+++ lib/web_ui/lib/ui.dart
@@ -32,6 +32,8 @@ part 'src/ui/initialization.dart';
 part 'src/ui/lerp.dart';
 part 'src/ui/natives.dart';
 part 'src/ui/painting.dart';
+part 'src/ui/path.dart';
+part 'src/ui/path_metrics.dart';
 part 'src/ui/pointer.dart';
 part 'src/ui/semantics.dart';
 part 'src/ui/test_embedding.dart';
